fix:목업 삭제
목업 삭제와 interface 위치 변경하였습니다.
컴포넌트와 utils 함수에서 원래 사용하던 interface의 위치를 변경된 위치로 갱신하였습니다.
