+++ src/state/productItemState.ts
@@ -1,7 +1,34 @@
 import { atom } from "recoil";
-import { Z_ASCII } from "zlib";
-import { ProductListValues } from "../mockup/productList";
 import { IOrderSelectedItem } from "../Page/Client/ClientMenu";
+
+export interface ProductOptions {
+  id: number;
+  name: string;
+}
+
+export interface ProductListValues {
+  id: number;
+  name: string;
+  price: number;
+  options?: ProductOptions[];
+  imageUrl?: string | null | undefined;
+  description?: string | null | undefined;
+  isAvailable?: boolean | undefined;
+}
+
+export interface Sales {
+  name: string;
+  price: number;
+  option: string;
+  quantity: number;
+  createdAt: string;
+}
+
+export interface SalesInfo {
+  productId: number;
+  createdAt: string;
+  sales: Sales[];
+}
 
 export enum Option {
   NONE = "none",
